Clean context module
